+++ src/preview/openscad-exe.ts
@@ -9,6 +9,9 @@ import { type } from 'os'; // node:os
 import { promisify } from 'util';
 
 import commandExists = require('command-exists');
+import { realpath } from 'fs/promises';
+import { ExtensionContext } from 'vscode';
+
 import { LoggingService } from 'src/logging-service';
 
 const execFile = promisify(child.execFile);
@@ -32,23 +35,34 @@ export class OpenscadExecutableManager {
     private openscadPath?: string;
     private arguments_: string[] = [];
 
-    public constructor(private loggingService: LoggingService) {}
+    public constructor(
+        private readonly loggingService: LoggingService,
+        private readonly context: ExtensionContext
+    ) {}
 
     private async getOpenscadVersion(
         openscadPath: string,
         arguments_: string[] = []
     ): Promise<string | undefined> {
-        const { stdout, stderr } = await execFile(openscadPath, [
-            ...arguments_,
-            '--version',
-        ]);
+        try {
+            const { stdout, stderr } = await execFile(
+                openscadPath,
+                [...arguments_, '--version'],
+                { cwd: this.context.extensionPath.toString() }
+            );
 
-        // For some reason, OpenSCAD seems to use stderr for all console output...
-        // If there is no error, assume stderr should be treated as stdout
-        // For more info. see: https://github.com/openscad/openscad/issues/3358
-        const output = stdout || stderr;
+            // For some reason, OpenSCAD seems to use stderr for all console output...
+            // If there is no error, assume stderr should be treated as stdout
+            // For more info. see: https://github.com/openscad/openscad/issues/3358
+            const output = stdout || stderr;
 
-        return output.trim().match(/version (\S+)/)?.[1];
+            return output.trim().match(/version (\S+)/)?.[1];
+        } catch (error) {
+            this.loggingService.logError(
+                `Error getting OpenSCAD version: ${error}`
+            );
+            return undefined;
+        }
     }
 
     /** Set the path to `openscad.exe` on the system.
@@ -57,7 +71,8 @@ export class OpenscadExecutableManager {
      */
     public async updateScadPath(
         newOpenscadPath?: string,
-        newArguments: string[] = []
+        newArguments: string[] = [],
+        skipLaunchPathValidation = false
     ): Promise<void> {
         if (
             newOpenscadPath === this.openscadPath &&
@@ -71,29 +86,65 @@ export class OpenscadExecutableManager {
         this.openscadExecutable = undefined;
 
         // Use platform default if not specified
-        const openscadPath = this.getPath();
+        let openscadPath = this.getPath();
 
-        this.loggingService.logInfo(`Using OpenSCAD path: '${openscadPath}'`);
+        this.loggingService.logInfo(
+            `Checking OpenSCAD path: '${openscadPath}'`
+        );
+
+        // Resolve potential symlinks
+        try {
+            const newPath = await realpath(openscadPath);
+            if (newPath !== openscadPath) {
+                this.loggingService.logInfo(
+                    `Configured path is a link. Using resolved path: '${openscadPath}'`
+                );
+                openscadPath = newPath;
+            }
+        } catch (error) {
+            // An ENOENT error (Error No Entity) is expected if openscadPath is
+            // invalid and is ok. Otherwise, throw the error
+            if (
+                !(error instanceof Error) ||
+                !('code' in error) ||
+                error.code !== 'ENOENT'
+            ) {
+                throw error;
+            }
+        }
 
         // TODO: Replace with something less nested
         commandExists(openscadPath, async (error: null, exists: boolean) => {
             if (!exists) {
-                return;
+                this.loggingService.logWarning(
+                    `'${openscadPath}' is not a valid path or command.`
+                );
+                if (!skipLaunchPathValidation) {
+                    return;
+                }
+                this.loggingService.logInfo('Skipping OpenSCAD path check.');
             }
-            const version = await this.getOpenscadVersion(
+            let version = await this.getOpenscadVersion(
                 openscadPath,
                 this.arguments_
             );
             // Should we throw an error here?
             if (!version) {
-                return;
+                this.loggingService.logWarning(
+                    `Unable to determine OpenSCAD version with 'openscad --version'.`
+                );
+                if (!skipLaunchPathValidation) {
+                    return;
+                }
+                this.loggingService.logInfo('Skipping OpenSCAD version check.');
+                version = 'unknown';
             }
             this.openscadExecutable = {
                 version: version,
                 filePath: openscadPath,
                 arguments_: this.arguments_,
             };
-            this.loggingService.logDebug(
+            this.loggingService.logInfo(
                 'Using OpenSCAD:',
                 this.openscadExecutable
             );